requirements.txt
Flask==2.0.3
flask-restx==0.5.1
pymongo==4.0.1
PyJWT==2.3.0